backend/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
requests==2.31.0
pydantic>=2.6.0
python-dotenv==1.0.0
httpx==0.26.0
itsdangerous==2.1.2